+++ app/Http/Controllers/Admin/ExpenseController.php
@@ -77,7 +77,7 @@ class ExpenseController extends Controller
 
     public function massDestroy(MassDestroyExpenseRequest $request)
     {
-        Expense::whereIn('id', request('ids'))->delete();
+        Expense::whereIn('id', request('ids'))->where('created_by_id', '=', auth()->user()->id)->delete();
 
         return response(null, Response::HTTP_NO_CONTENT);
     }
