+++ app/Http/Controllers/Admin/IncomeController.php
@@ -77,7 +77,7 @@ class IncomeController extends Controller
 
     public function massDestroy(MassDestroyIncomeRequest $request)
     {
-        Income::whereIn('id', request('ids'))->delete();
+        Income::whereIn('id', request('ids'))->where('created_by_id', '=', auth()->user()->id)->delete();
 
         return response(null, Response::HTTP_NO_CONTENT);
     }
